Add python bindings for hwlib (#21)
* Add python bindings for hwlib

* Fix readme

* Remove _py suffix from python bindings

* Return python dict instead of string

* Fix style issues

* Add tests for rust code

* Run cargo test in the GH actions

* Add copyright message to each rust file

* Add pytest check to CI

* Fix CI

* Update dependencies

* Run pre-commit hooks only if server files have been modified

* Update client action

* Update readme

* Fix virtualenv in CI

* Update CI

* Create global README

* Use async in python func

* Run py tests with tox

* s/use/build

* Update links in README

+++ scripts/fetch_openapi_yaml.sh
@@ -1,18 +1,21 @@
 #!/bin/bash
 
-cd server
+# Run only if server code has been modified
+if git diff --cached --name-only | grep --quiet "server"
+then
+    cd server
+    # Start FastAPI app in the background
+    poetry run uvicorn hwapi.main:app --port 8002 &
 
-# Start FastAPI app in the background
-poetry run uvicorn hwapi.main:app --port 8002 &
+    # Wait for the server to start
+    sleep 5
 
-# Wait for the server to start
-sleep 5
+    # Fetch OpenAPI YAML
+    curl http://localhost:8002/v1/openapi.yaml -o schemas/openapi.yaml
 
-# Fetch OpenAPI YAML
-curl http://localhost:8002/v1/openapi.yaml -o schemas/openapi.yaml
+    # Kill the FastAPI server
+    kill $!
 
-# Kill the FastAPI server
-kill $!
-
-# Add schema to the committed files
-git add schemas/openapi.yaml
+    # Add schema to the committed files
+    git add schemas/openapi.yaml
+fi
